change migration

+++ database/migrations/2023_11_22_150952_create_sponsors_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -15,10 +14,10 @@ return new class extends Migration
     {
         Schema::create('sponsors', function (Blueprint $table) {
             $table->id();
-            
-            $table->double('price');
+
+            $table->decimal('price');
             $table->integer('time');
-            $table->string('name',50);
+            $table->string('name', 50);
 
             $table->timestamps();
         });
